r250/sofa-dev : Update qmake to version 4.1.2 for windows. With this version projects files and generated for both Debug and Release configurations, and support for Visual Studio 2005 is added. Also the sofa.cfg is no longer modified depending on the version of Visual Studio in use. Still need to recompile fltk/libxml2/iconv to remove warning when linking with VC8.
 //partial sync (8/115 changes)//

+++ src/Sofa/Components/Common/RigidTypes.h
@@ -158,9 +158,9 @@ public:
         void writeOpenGlMatrix( float m[16] ) const
         {
             orientation.writeOpenGlMatrix(m);
-            m[12] = center[0];
-            m[13] = center[1];
-            m[14] = center[2];
+            m[12] = (float)center[0];
+            m[13] = (float)center[1];
+            m[14] = (float)center[2];
         }
 
         /// compute the projection of a vector from the parent frame to the child
